Conflicts:
	.gitignore
	nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.infra/src/main/java/nts/uk/ctx/bs/employee/infra/repository/temporaryabsence/JpaTemporaryAbsence.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/optitem/KrcstCalcResultRange_.java
@@ -32,10 +32,10 @@ public class KrcstCalcResultRange_ {
 	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> lowerTimeRange;
 
 	/** The upper number range. */
-	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> upperNumberRange;
+	public static volatile SingularAttribute<KrcstCalcResultRange, Double> upperNumberRange;
 
 	/** The lower number range. */
-	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> lowerNumberRange;
+	public static volatile SingularAttribute<KrcstCalcResultRange, Double> lowerNumberRange;
 
 	/** The upper amount range. */
 	public static volatile SingularAttribute<KrcstCalcResultRange, Integer> upperAmountRange;
